refactor: organize project structure and add auth context
