+++ estoque-vendas/jest.config.js
@@ -8,7 +8,8 @@ const customJestConfig = {
   moduleNameMapper: {
     '^@/(.*)$': '<rootDir>/src/$1',
   },
-  testEnvironment: 'node',
+  testEnvironment: 'jsdom',
+  setupFilesAfterEnv: ['<rootDir>/tests/setup.js'],
 };
 
 module.exports = createJestConfig(customJestConfig);
