+++ estoque-vendas/next.config.js
@@ -1,0 +1,12 @@
+const path = require('path');
+
+/** @type {import('next').NextConfig} */
+const nextConfig = {
+  reactStrictMode: true,
+  webpack: (config) => {
+    config.resolve.alias['@'] = path.join(__dirname, 'src');
+    return config;
+  },
+};
+
+module.exports = nextConfig;
