+++ estoque-vendas/src/pages/_app.js
@@ -1,5 +1,10 @@
-import "@/styles/globals.css";
+import '@/styles/globals.css';
+import { AuthProvider } from '@/context/AuthContext';
 
 export default function App({ Component, pageProps }) {
-  return <Component {...pageProps} />;
+  return (
+    <AuthProvider>
+      <Component {...pageProps} />
+    </AuthProvider>
+  );
 }
